re-implemented select/deselect all checkboxes and naming-override checkbox
