--- Hardware/scripts/cse_manager/src/.env
@@ -1,1 +1,0 @@
-INPUT_ZIP_FOLDER=library_input
