+++ Hardware/scripts/kicad_library_manager/src/.env
@@ -1,0 +1,1 @@
+INPUT_ZIP_FOLDER=library_input
